Fixed the major bug in activating the camera within the profile page.

+++ app/src/main/java/com/example/connectue/ProfileFragment.java
@@ -344,17 +344,17 @@ public class ProfileFragment extends Fragment {
             firstNameStr = (String) document.get(General.FIRSTNAME);
             Log.i(TAG,"after: " + firstNameStr);
         }
-        if (document.get(General.LASTNAME) != null) {  lastNameStr = (String) document.get(General.LASTNAME);}
-        if (document.get(General.EMAIL) != null) { emailStr = (String) document.get(General.EMAIL);}
-        if (document.get(General.PROGRAM) != null) { majorStr = (String) document.get(General.PROGRAM);}
-        if (document.get(General.PHONE) != null) { phoneStr = (String) document.get(General.PHONE);}
+        if (document.get(General.LASTNAME) != null) {  lastNameStr = document.getString(General.LASTNAME);}
+        if (document.get(General.EMAIL) != null) { emailStr = document.getString(General.EMAIL);}
+        if (document.get(General.PROGRAM) != null) { majorStr = document.getString(General.PROGRAM);}
+        if (document.get(General.PHONE) != null) { phoneStr = document.getString(General.PHONE);}
         if (document.get(General.ROLE) != null) {
             role = document.getLong(General.ROLE);
 //            isAdmin = General.isAdmin(document.getLong(General.ROLE));
 //            Log.i(TAG, "" + isAdmin);
         }
         if (document.get(PROFILEPICTURE) != null) {
-            imageURL = (String) document.get(PROFILEPICTURE);
+            imageURL = document.getString(PROFILEPICTURE);
         }
     }
 
@@ -447,6 +447,7 @@ public class ProfileFragment extends Fragment {
 
     private void captureImageFromCamera() {
         //intent to pick image from camera
+        Log.i(TAG_Profile, "captureImageFromCamera entered ");
         ContentValues cv = new ContentValues();
         cv.put(MediaStore.Images.Media.TITLE, "Temp Pick");
         cv.put(MediaStore.Images.Media.DESCRIPTION, "Temp Descr");
@@ -455,6 +456,21 @@ public class ProfileFragment extends Fragment {
         Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
         intent.putExtra(MediaStore.EXTRA_OUTPUT, imageUri);
         startActivityForResult(intent, IMAGE_PICK_CAMERA_CODE);
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        Log.i(TAG_Profile, "activity result entered ");
+        if (resultCode == Activity.RESULT_OK) {
+            if (requestCode == IMAGE_PICK_GALLERY_CODE) {
+                //image picked from gallery
+                imageUri = data.getData();
+            }
+            profileIV.setImageURI(imageUri);
+            profileIV.setVisibility(View.VISIBLE);
+            updateProfilePicture();
+        }
     }
 
     @Override
@@ -479,30 +495,13 @@ public class ProfileFragment extends Fragment {
         }
     }
 
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (resultCode == Activity.RESULT_OK) {
-            if (requestCode == IMAGE_PICK_GALLERY_CODE) {
-                //image picked from gallery
-                imageUri = data.getData();
-            }
-            profileIV.setImageURI(imageUri);
-            profileIV.setVisibility(View.VISIBLE);
-            updateProfilePicture();
-        }
-    }
-
     private void updateProfilePicture() {
-
+        Log.i(TAG_Profile, "updateProfilePicture entered");
         if (imageUri == null) {
             return;
         }
 
-        StorageReference filePath = FirebaseStorage.getInstance().getReference("posts")
-                .child(emailStr + "." + General.getFileExtension(getContext(), imageUri));
-
-        if (imageURL != null || !imageURL.equals("")) {
+        if (imageURL != null && !imageURL.equals("")) {
             StorageReference currentPicture = FirebaseStorage.getInstance().
                     getReferenceFromUrl(imageURL);
             currentPicture.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
@@ -518,6 +517,9 @@ public class ProfileFragment extends Fragment {
                         }
                     });
         }
+
+        StorageReference filePath = FirebaseStorage.getInstance().getReference("profilePicture")
+                .child(emailStr + "." + General.getFileExtension(getContext(), imageUri));
 
         filePath.putFile(imageUri)
                 .addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
@@ -539,8 +541,13 @@ public class ProfileFragment extends Fragment {
                                     }
                                 });
                     }
+                }).addOnFailureListener(new OnFailureListener() {
+                    @Override
+                    public void onFailure(@NonNull Exception e) {
+                        Log.i(TAG_Profile, "In updateProfilePicture, upload to storage failed");
+                    }
                 });
-
+        Log.i(TAG_Profile, "In updateProfilePicture, after upload");
     }
 
 }
